squash: doesn't work
There's a whole bunch of ugly code here in the solver.  In theory there is
everything needed except the code that automatically moves a card from
a column or one of the cells to the foundation.  However, the backtracking
isn't working, so a few things are tried, but nothing that ever exposes a
needed card, so we never even hit that code.

This is just an intermediate commit so I have something to fall back on
in case I get a bee up my butt and decide to refactor instead of looking
at the backtracking failure.