squash: backtracking seems better now
No guarantee that it's actually working.  I'm surprised at the depth it gets
to before it needs to move a card to the foundation.  That could be a sign
that my seen / hash stuff is not working properly.  It could also be a sign
that I don't know what I'm doing ...

Anyway, I need to write the code that allows a card to be moved to the
foundation.  I'll probably do some refactoring before hand because there's
an awful lot of goofy calculation of ranks and/or suits that could trivially
be sped up.

Now's a good stopping point for the night, although it's a bit disappointing
that this will have taken me more than a day to write.  OTOH, I did get
interrupted by (paid) work and family ...