squash: some quick code before I go off to run.
I'm now at a good point to properly implement automatic moves.  If I did
everything correctly, that will give me the solver.  However, the solver
is remembering all the boards, so it might run out of memory.  OTOH, we
really only need to remember all the hashes, because they're perfect and
that will allow a lot of garbage collection.