Oops!  Allow moves from a cell to an empty column.
The previous version inadvertently had an extra constraing which was that
a card from one of the cells could only be moved down to a column which
already had at least one card.  That was not deliberate.

This also gets rid of LOAD_PATH adjustment, which makes it easier to pull
the solver or board in with irb -r