Beginning of image reading portion of the code
This has code that successfully reads ranks and suits from an image.
It prints hashes for each, so I should be able to create a little
hash map to help us out.  Ranks are always located relative to
a fixed spot.  Suits wiggle around a little, so we need to look
for white columns and non-white rows to figure out exactly where
to calculate the hash from.

This code can now be used to write a function for turning an image
into board.