contains some cleanups.
This is a bit of a fool's errand.  I hadn't looked at this code in a long
time, so I decided to fix the clippy warnings, then fix TODOs and other
comments, then started doing little cleanups here and there.

I'm doing this in part because I'm writing a presentation about learning
Rust and this is some of my oldest Rust code...